+++ App/controllers/student.py
@@ -1,15 +1,24 @@
 from App.models import Student
 from App.database import db
+import json
 
-def create_student(name):
-    newStudent= Student(name=name)
+def create_student(studentID, name):
+    student=Student.query.filter_by(studentID=studentID).first()
+    if student:
+        return None
+    
+    # if student has not already been added
+    newStudent= Student(studentID, name=name)
     db.session.add(newStudent)
     db.session.commit()
+    return newStudent
 
 def delete_student(studentID):
     student=Student.query.get(studentID)
-    db.session.delete(student)
-    db.session.commit()
+    if student:
+        db.session.delete(student)
+        db.session.commit()
+    return student
 
 def update_student(studentID, name):
     student=Student.query.filter_by(studentID=studentID).first()
@@ -20,7 +29,7 @@ def update_student(studentID, name):
     return student
 
 def get_student(studentID):
-    return Student.query.get(studentID)
+    return Student.query.filter_by(studentID=studentID).first()
 
 def get_all_students():
     return Student.query.all()
@@ -29,8 +38,9 @@ def get_all_students_JSON():
     students=get_all_students()
     if not students:
         return []
-    students = [student.toJSON() for student in students]
-    return students
+    else:
+        students = [student.toJSON() for student in students]
+        return students
